Make ActionCreators work with es6 arrow functions
:lipstick:

ES6 arrow functions are bound to the current scope making it impossible to call `this.dispatch(...)` or other methods of the ActionCreators inside the callback. 

By passing context, one can use `context.dispatch(...)` or other methods of the ActionCreator.

Since context is passed as the last argument, all existing code work as is and if you don't use es6 arrow syntax, you're not affected.